refactor(commenting): split comments-overlay into focused modules (#9796)
In order to keep working on the commenting layer without loading ~1500
lines of unrelated context every time, this PR splits
`comments-overlay.tsx` into modules along the seams that were already
there. The file that landed in #9471 holds the layer, the pin, the
cluster badge, the region box and its resize handles, the pending
composer, the fade reconciler, and the clustering state machine — seven
independent concerns in one file, which makes every follow-up (mobile
placement, composability, region work) a merge risk against all the
others.

Nothing about the rendered output changes. Six of the seven pieces move
verbatim — the only edit to their code is the added `export`. The one
substantive change is the clustering state: lines 219–368 of the layer
were a contiguous run of hooks plus render-phase state transitions
maintaining the rendered-vs-latest partition, and they become a
`useClusterModel` hook returning `{ model, zoomBounds, fadeNodes,
orphanThreads, heldThreads }`. Hook order is preserved exactly, and the
code inside is byte-for-byte identical — the only difference is that the
invariant comment ("the only thing that re-flows clustering doc-wide is
zoom") now sits on the hook that enforces it rather than buried
mid-component.

| Module | Lines | Owns |
| --- | --- | --- |
| `comments-overlay.tsx` | 412 | `CanvasComments`, render composition,
portal host, shortcuts |
| `thread-pin.tsx` | 410 | `ThreadPin` — marker, drag-to-reanchor,
popover |
| `cluster-model.ts` | 344 | `useClusterModel`, `revealThreadPin`,
`zoomToClusterSplit` |
| `pending-composer.tsx` | 152 | `PendingComposer` |
| `region-box.tsx` | 124 | `RegionBox`, `RegionResizeHandles`, corner
geometry |
| `cluster-badge.tsx` | 124 | `ClusterBadge`, viewport culling |
| `cluster-fade.ts` | 102 | enter/present/exit fade reconciler |
| `canvas-events.ts` | 18 | `isCanvasPanGesture`,
`forwardPointerEventToCanvas` |

Two notes for review:

- `useFadeVisibleNodes` takes its `resetKey` as `object` rather than the
`ClusterModel` shape it's always passed. It only ever compares by
identity, and the looser type keeps `cluster-fade` from importing
`cluster-model` while `cluster-model` imports it back.
- `zoomToClusterSplit` moved out of the layer as a plain function rather
than staying a `useCallback` there, because it shares
`CLUSTER_SPLIT_ZOOM_FACTOR` and `centerOnPointAtZoom` with
`revealThreadPin`; leaving it behind would mean exporting both of those
instead.

There's no public API change — `CanvasComments` and
`CanvasCommentsProps` still export from `comments-overlay.tsx`, and
nothing else in the repo imported from that file.

Unrelated, spotted while splitting and deliberately left alone:
`.tlui-cmt-canvas-region--movable` is dead. No component sets it, so the
CSS rule in `canvas.css` and the check in the pin's dismiss handler both
reference a class that can't appear — a leftover from a removed "movable
region body" mode. Worth its own cleanup.

### Change type

- [x] `other`

### Test plan

1. Open a board with several comments and zoom out until pins fold into
cluster badges; zoom back in and confirm they split and cross-fade as
before.
2. Drag a pin while it's folded into a badge — it should pop out and
ride its anchor, then rejoin on the next zoom-out.
3. Drag a pin onto a shape and confirm it re-anchors, with the
drop-target hint showing mid-drag.
4. Create a region comment, resize it from a corner, and drag it by its
pin.
5. Open a thread, dismiss with Escape and with a click away; toggle pins
with Shift+C.
6. Click a thread in the sidebar and confirm the reveal zooms to and
opens its pin.

- [ ] Unit tests
- [ ] End to end tests

No new tests — this is a move. The existing 281 tests in
`packages/commenting` pass unchanged, as do `yarn typecheck` and `yarn
api-check`.

### Code changes

| Section   | LOC change   |
| --------- | ------------ |
| Core code | +1302 / -1139 |
